# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Codoctor API",
    "description": "Graduatio Project"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "local server"
    },
    {
      "url": "https://codoctor.onrender.com",
      "description": "remote server"
    }
  ],
...
  "paths": {
...
    "/api/doctors/": {
      "get": {
        "tags": [
          "Doctor"
        ],
        "description": "search for doctors by name or by specializtion",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
...
    "/api/doctors/{id}": {
      "get": {
        "tags": [
          "Doctor"
        ],
        "description": "Get doctor profile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {}
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}